+++ src/scripts/api.js
@@ -1,0 +1,8 @@
+import { EasyExport } from "./EasyExports.js";
+import CONSTANTS from "./constants.js";
+
+const API = {
+  easyExport: new EasyExport(),
+};
+
+export default API;