+++ src/scripts/settings.js
@@ -1,0 +1,63 @@
+import { debug, log, warn, i18n } from "./lib/lib.js";
+import CONSTANTS from "./constants.js";
+
+export const registerSettings = function () {
+  //   game.settings.registerMenu(CONSTANTS.MODULE_ID, "resetAllSettings", {
+  //     name: `${CONSTANTS.MODULE_ID}.setting.reset.name`,
+  //     hint: `${CONSTANTS.MODULE_ID}.setting.reset.hint`,
+  //     icon: "fas fa-coins",
+  //     type: ResetSettingsDialog,
+  //     restricted: true,
+  //   });
+
+  // ========================================================================
+
+  game.settings.register(CONSTANTS.MODULE_ID, "debug", {
+    name: `${CONSTANTS.MODULE_ID}.setting.debug.name`,
+    hint: `${CONSTANTS.MODULE_ID}.setting.debug.hint`,
+    scope: "client",
+    config: true,
+    default: false,
+    type: Boolean,
+  });
+};
+
+class ResetSettingsDialog extends FormApplication {
+  constructor(...args) {
+    //@ts-ignore
+    super(...args);
+    //@ts-ignore
+    return new Dialog({
+      title: game.i18n.localize(`${CONSTANTS.MODULE_ID}.dialogs.resetsettings.title`),
+      content:
+        '<p style="margin-bottom:1rem;">' +
+        game.i18n.localize(`${CONSTANTS.MODULE_ID}.dialogs.resetsettings.content`) +
+        "</p>",
+      buttons: {
+        confirm: {
+          icon: '<i class="fas fa-check"></i>',
+          label: game.i18n.localize(`${CONSTANTS.MODULE_ID}.dialogs.resetsettings.confirm`),
+          callback: async () => {
+            const worldSettings = game.settings.storage
+              ?.get("world")
+              ?.filter((setting) => setting.key.startsWith(`${CONSTANTS.MODULE_ID}.`));
+            for (let setting of worldSettings) {
+              log(`Reset setting '${setting.key}'`);
+              await setting.delete();
+            }
+            //window.location.reload();
+          },
+        },
+        cancel: {
+          icon: '<i class="fas fa-times"></i>',
+          label: game.i18n.localize(`${CONSTANTS.MODULE_ID}.dialogs.resetsettings.cancel`),
+        },
+      },
+      default: "cancel",
+    });
+  }
+
+  async _updateObject(event, formData = undefined) {
+    // do nothing
+  }
+}