+++ vite.config.mjs
@@ -17,7 +17,7 @@ import { run } from 'vite-plugin-run'
 
 // For convenience, you just need to modify the package ID below as it is used to fill in default proxy settings for
 // the dev server.
-const s_MODULE_ID = "Border-Control";
+const s_MODULE_ID = "easy-exports";
 const s_PACKAGE_ID = "modules/"+s_MODULE_ID;
 const s_ENTRY_JAVASCRIPT = "module.js";
 
@@ -61,8 +61,8 @@ export default () => {
 
     css: {
       // Creates a standard configuration for PostCSS with autoprefixer & postcss-preset-env.
-      postcss: postcssConfig({ 
-        compress: s_COMPRESS, 
+      postcss: postcssConfig({
+        compress: s_COMPRESS,
         sourceMap: s_SOURCEMAPS
       }),
     },
@@ -92,7 +92,7 @@ export default () => {
         "/socket.io": { target: "ws://127.0.0.1:30000", ws: true },
       },
     },
-    
+
     build: {
       outDir: normalizePath( path.resolve(__dirname, `./dist/${s_MODULE_ID}`)), // __dirname,
       emptyOutDir: false,
@@ -177,14 +177,14 @@ export default () => {
           if (warning.message.includes(`<a> element should have an href attribute`)) {
             return;
           }
-          
+
           // Let Rollup handle all other warnings normally.
           handler(warning);
         },
       }),
 
       resolve(s_RESOLVE_CONFIG), // Necessary when bundling npm-linked packages.
-      
+
       // When s_TYPHONJS_MODULE_LIB is true transpile against the Foundry module version of TRL.
       s_TYPHONJS_MODULE_LIB && typhonjsRuntime(),