+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # RELEASE NOTES
+## 0.6.5b
+
+- Fix use of entity (should be type) in metadata on import
+
 ## 0.6.5
 - Fixed: Issue [#11](https://github.com/spetzel2020/easy-exports/issues/11) ; use documentName not .entity
 - Because of this change, will only work with the Foundry 0.8+ data model
@@ -8,7 +12,7 @@
 - Test for Foundry 0.9
 - Fix deprecation of Directory.entities
 ## 0.6
-- Preserve folders and permissions to test whether these can be used on re-import from a Compendium      
+- Preserve folders and permissions to test whether these can be used on re-import from a Compendium
 - Check version and use new calls for Foundry 0.8
 ## 0.5.1 21-Dec-2020
 - Fixed: Full Backup was exporting/importing "null" entity which meant the world would not start up
@@ -35,7 +39,7 @@
 - Change logging to be more informative
 - Removed .setup()
 
-## 0.2.5    
+## 0.2.5
 Remove Compendium from README
 Verify compatibility with Foundry 0.7.2